ml-backend/requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
numpy
python-dotenv
httpx
pdfplumber
psycopg2-binary
prisma